medical-examination-assistant/diarization-service/requirements.txt
# Minimal dependencies for Pyannote diarization
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pyannote.audio>=3.1.0
torch>=2.0.0
torchaudio>=2.0.0
numpy>=1.24.0
soundfile>=0.12.0
